docs: split bigframes query samples into separate files
